% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biodivTools-package.R
\docType{package}
\name{biodivTools-package}
\alias{biodivTools-package}
\alias{biodivToolspackage}
\title{Tools for biodiversity analysis}
\description{
Tools for biodiversity analysis
}
\details{
We provide a set of functions usefull for some of our common biodviersity
analysis studies.
}
\author{
Thomas Nauss
}
\keyword{package}

