% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compSpecRichBExpVegReleves.R
\name{compSpecRichBExpVegReleves}
\alias{compSpecRichBExpVegReleves}
\title{Compute species richness from vegetation releves data}
\usage{
compSpecRichBExpVegReleves(vegrel)
}
\arguments{
\item{vegrel}{\code{data.frame} containing multi-year vegetation releves data.}
}
\value{
A \code{data.frame} object.
}
\description{
Compute species richness from multi-year vegetation releves data in the
format of the DFG Biodiversity Exploratories.
}
\examples{
\\notrun{
}

}
\references{
NONE
}
\author{
Thomas Nauss
}
