% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBExpLUI.R
\name{readBExpLUI}
\alias{readBExpLUI}
\title{Read LUI data}
\usage{
readBExpLUI(filepath, ...)
}
\arguments{
\item{filepath}{Full path and name of the csv file containing the LUI data.}
}
\value{
A \code{data.frame} object.
}
\description{
Read land use intensity index data from files provided in the format of the
DFG Biodiversity Exploratories.
}
\examples{
\\notrun{
}

}
\references{
NONE
}
\author{
Thomas Nauss
}
