% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBExpSMI.R
\name{readBExpSMI}
\alias{readBExpSMI}
\title{Read SMI data}
\usage{
readBExpSMI(filepath, ...)
}
\arguments{
\item{filepath}{Full path and name of the csv file containing the SMI data.}
}
\value{
A \code{data.frame} object.
}
\description{
Read forest intensity index data from files provided in the format of the
DFG Biodiversity Exploratories.
}
\examples{
\\notrun{
}

}
\references{
NONE
}
\author{
Thomas Nauss
}
