% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBExpStandStruc.R
\name{readBExpStandStruc}
\alias{readBExpStandStruc}
\title{Read stand structural attribute data.}
\usage{
readBExpStandStruc(filepath)
}
\arguments{
\item{filepath}{Full path and name of the stand structural file.}
}
\value{
A \code{data.frame} object.
}
\description{
Read stand structural attribute data from files provided in the format of the
DFG Biodiversity Exploratories.
}
\examples{
\\notrun{
}

}
\references{
NONE
}
\author{
Thomas Nauss
}
