% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBExpVegHeaderData.R
\name{readBExpVegHeaderData}
\alias{readBExpVegHeaderData}
\title{Read vegetation releves data}
\usage{
readBExpVegHeaderData(filepath)
}
\arguments{
\item{vegrel}{Full path and name of the vegetation releves file.}
}
\value{
A \code{data.frame} object.
}
\description{
Read vegetation releves data from files provided in the format of the
DFG Biodiversity Exploratories.
}
\examples{
\\notrun{
}

}
\references{
NONE
}
\author{
Thomas Nauss
}
