% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBExpPoly.R
\name{readBExpPoly}
\alias{readBExpPoly}
\title{Read plot polygon data}
\usage{
readBExpPoly(shp, crs, ...)
}
\arguments{
\item{shp}{Full path and name of the shp file.}
}
\value{
A \code{SpatialPolygonDataFrame} object.
}
\description{
Read plot polygons in the format of the DFG Biodiversity Exploratories and
adjust the plot IDs.
}
\examples{
\\notrun{
}

}
\author{
Thomas Nauss
}
\references{
NONE
}

